pyproject.toml
[project]
name = "parrandatathon2025"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langgraph>=0.3.25",
    "openai>=1.12.0",
    "python-dotenv>=1.0.0",
    "pandas>=2.2.0",
    "pyarrow>=15.0.0",
    "rank-bm25>=0.2.2",
    "h5py>=3.13.0",
    "langchain-openai>=0.3.12",
    "chromadb>=1.0.0",
    "aiohttp>=3.11.16",
    "nltk>=3.9.1",
    "transformers>=4.50.3",
    "torch>=2.6.0",
    "scikit-learn>=1.6.1",
    "numpy>=1.26.4",
    "fuzzywuzzy>=0.18.0",
    "python-levenshtein>=0.27.1",
    "llama-stack>=0.2.0",
    "faiss-cpu>=1.10.0",
    "fastapi>=0.115.9",
    "uvicorn>=0.34.0",
    "pydantic>=2.11.2",
    "asyncio>=3.4.3",
    "ipython>=9.0.2",
]
